Collapse writes_validation into writes.validate(); fix conftest cycle
writes_validation.py existed only to dodge a cycle that no longer exists
once the validation call was moved out of projections/__init__.py. Move
ConfigurationError + the cross-registry assertion back into writes.py as
writes.validate(). writes.py already imports from projections (it uses
WorkerAttrsProjection), so adding the PROJECTIONS import at top level is
direction-preserving — no cycle.

Also revert one bulk-lift in tests/cluster/controller/conftest.py: the
test conftests have a mutual dependency (cluster/conftest imports
make_test_entrypoint from cluster/controller/conftest; the controller one
needs fake_log_client_from_service from cluster/conftest), so keep the
latter as a function-local import. The static-cycle script does not scan
test directories, which is why the regression slipped through.
